requirements.txt
torch
numpy
matplotlib
pandas
scipy
tqdm
progressbar2
coloredlogs
tensorboard
pyyaml
tables
